Fix concurrent memoization of zero-arg methods multiple values
When two threads concurrently call a zero-arg method which is in
the unmemoized state, they can return two different results.

For example, this could occur if we are memoizing:
  * database query result
  * filesystem read
  * current time
  * a random number

This PR introduces a test which probabilistically reproduces the
issue, as well as a fix which use a "double check" pattern to
store the result only if a competing thread has not already won
the race.